requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7                # via pip-tools
decorator==4.2.1          # via ipython, traitlets
django-environ==0.4.4
django-extensions==2.0.0
django==2.0.2
djangorestframework==3.7.7
first==2.0.1              # via pip-tools
ipython-genutils==0.2.0   # via traitlets
ipython==6.2.1
jedi==0.11.1              # via ipython
markdown==2.6.11
parso==0.1.1              # via jedi
pexpect==4.4.0            # via ipython
pickleshare==0.7.4        # via ipython
pip-tools==1.11
prompt-toolkit==1.0.15    # via ipython
psycopg2==2.7.4
ptyprocess==0.5.2         # via pexpect
pygments==2.2.0           # via ipython
pytz==2018.3              # via django
pyyaml==3.12
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via django-environ, django-extensions, pip-tools, prompt-toolkit, traitlets
traitlets==4.3.2          # via ipython
typing==3.6.4             # via django-extensions
wcwidth==0.1.7            # via prompt-toolkit
